Consolidate make_date
